update react code
