+++ frontend/src/App.css
@@ -1,44 +1,17 @@
-/* .App {
-  text-align: center;
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-} */
-
 .App{
   display: flex;
 }
 
 .LeftsideBar{
+  background-color: #98e9dc;
+  color: #666;
   height: 100vh;
+  position: fixed;
   width: 10vw;
+}
+
+.LeftsideBar a:visited{
+  color: #666;
 }
 
 .LeftsideBar ul{
@@ -48,6 +21,7 @@
 }
 
 .LeftsideBar li{
-  outline: solid 1px #555;
+  text-align: center;
+  /* outline: solid 1px #555; */
   padding: 10px 0px 10px 0px;
 }
