+++ frontend/src/SubjectPosts.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class SubjectPosts extends React.Component {
+    render() {
+        return (
+            <div className="SubjectPosts">
+                <h1>{this.props.subjectsName}</h1>
+            </div>
+        )
+    }
+}
+
+export default SubjectPosts;
