+++ frontend/src/Zwitter.js
@@ -1,4 +1,7 @@
 import React from 'react';
+import './Zwitter.css';
+
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import { BrowserRouter as Router, Link } from 'react-router-dom';
 
@@ -6,7 +9,12 @@ class Zwitter extends React.Component {
   render() {
     return (
       <div className="Zwitter">
-        <h1>This is Zwitter.</h1>
+        <div className="ZwitterContents">
+          <h1>This is Zwitter.</h1>
+        </div>
+        <div className="RightsideBar">
+          <FontAwesomeIcon icon={['fas', 'cog']} size="5x" className="configIcon" />
+        </div>
       </div>
     )
   }
